+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-11862 Added polling for dir stats status when disabled
 * VFS-12053 Removed OpenData emblem from public share description view
 * VFS-11983 Enabled EDM editor for newly created Open Data metadata
 * VFS-12042 Fixed always disabled share actions when accessed via Shares main menu